HTML DOM and Play/Pause
Fixed an error with the HTML doc DOM thing, not sure why it popped up
now. Also have the play pause functioning so that once resuming the
animation the block starts where it left off and doesn't "jump" to where
it would have beed had the animation not been paused.